API Mistakes
Rect.new() could be constructed with vectors, if the first parameter is a Vector2 then it REQUIRES the second parameter to be defined. Not sure how to show that.
UDim2.new() doesn't take any as the first two constructors, it's two UDims.
BrickColor.new() has diff params.
CFrame.new() has diff params.