Print linting summary as JSON if display-style is json or json2
When requesting "display-style" to be json(2), optimally the entire
output of the application should be JSON, so that it can be easily
consumed by another application, without having to filter out
non-JSON lines. This PR adds the ability to print the summary as JSON
instead of printing it as plain text.