Add react lint (#521)
Closes #484.

The alternative implementation is to bundle React logic in the existing
`roblox_incorrect_roact_usage` lint to reduce duplicated code. I decided
to make a new lint for React so separating roact and react-lua logic is
more sane as more rules specific to react-lua is added. But it's an easy
fix if we want to merge the React rules into the existing roact lint.

Note the layout in the doc. Since this includes 3 rules, I found
organizing this by rules is more readable, but it's also an easy change
if we prefer consistency with other existing docs.