Show missing required param message when in standard library format (#545)
Closes #544.

If the function has multiple required parameters each with their own
message, this will only show the first missing param if it has a
message.

The alternative to this is to show every missing parameter's message
with an indicator to which arg they associate with, but this may be
confusing in cases like `assert()` with no args. The error message would
show the standard library format message of `A failed assertion without
a message is unhelpful to users`, but the bigger issue is they aren't
asserting anything at all.